[gstreamer] make sure we process all frames when we get EOS. Fixes JB#8433
The problem is that we keep a reference to the input buffer to avoid destroying
it before hw encoder is done with it.

If we get an EOS while the encoder still has a queue of buffers and we stop the
src pad task then we will never release those buffers back to the upstream element
when encoder signals us (EmptyBufferdone).

What we do is to keep the task alive until all buffers have been processed and then
we can safely stop it.