[gstreamer] Don't enqueue non-android buffers to the codec after unflushing the port. Fixes JB#8942
Seems it's not needed for encoders. What happens is we get a full buffer from
the encoder already. If we pass the buffers back then we simply lose the buffer.

A better idea would be to mark any encoder returned buffer and not hand it back
to the encoder but the whole step is not really needed at least for qualcomm encoders.