[gstreamer] properly pass buffers to omx component output port. Fixes JB#8942
What happens is we have 1 queue which holds the buffers to be passed to omx and
the buffers returned from omx after filling.

We initially start popping from the queue and passing the buffers to omx when we
unflush the port. omx takes the first buffer, fills it with codec config, passes
it again to our queue which we are still popping and thus we end up eating the
buffer width codec config by mistake.

We cannot really use locking here because FillBufferdone could be invoked from the
same thread and we end up deadlocking.

The solution I did was not to pass the buffers back to omx initially.
This worked fine but lead to the output handling function picking up
those empty buffers and trying to handle them as output buffers.

A better solution is to copy the queue before we start unflushing,
clear the queue that is used by FillBufferDone and output handling
function and then popping and passing buffers from the new queue.
If we get buffers from the encoder then they will be handled properly and not eaten.