[gst-omx] Fix timeout when reconfiguring output port. Contributes to JB#17654
Ensure buffers acquired in the decoder loop are returned to the pending
queue if they are not pushed for any reason, otherwise they will leak
and won't be available to free on de-allocation.

Don't wait for pushed buffers to be returned to the decoder before
freeing their buffer header.  Instead simply invalidate them and let
the reference counting take care of deletion when nothing else
refers to them.